Moved hystersis to core.utils
